[#2961] Fix broken data preview test and move to own file
