+++ ckan/tests/lib/test_helpers.py
@@ -177,21 +177,3 @@ class TestHelpers(TestController):
         assert_equal(h.get_pkg_dict_extra(pkg_dict, 'extra_not_found'), None)
 
         assert_equal(h.get_pkg_dict_extra(pkg_dict, 'extra_not_found', 'default_value'), 'default_value')
-
-    def test_compare_domains(self):
-        ''' see https://en.wikipedia.org/wiki/Same_origin_policy
-        '''
-        comp = h._compare_domains
-        assert comp(['http://www.okfn.org', 'http://www.okfn.org']) == True
-        assert comp(['http://www.okfn.org', 'http://www.okfn.org', 'http://www.okfn.org']) == True
-        assert comp(['http://www.OKFN.org', 'http://www.okfn.org', 'http://www.okfn.org/test/foo.html']) == True
-        assert comp(['http://okfn.org', 'http://okfn.org']) == True
-        assert comp(['www.okfn.org', 'http://www.okfn.org']) == True
-        assert comp(['//www.okfn.org', 'http://www.okfn.org']) == True
-
-        assert comp(['http://www.okfn.org', 'https://www.okfn.org']) == False
-        assert comp(['http://www.okfn.org:80', 'http://www.okfn.org:81']) == False
-        assert comp(['http://www.okfn.org', 'http://www.okfn.de']) == False
-        assert comp(['http://de.okfn.org', 'http://www.okfn.org']) == False
-
-        assert comp(['http://de.okfn.org', 'http:www.foo.com']) == False
